+++ packages/web3-games/src/games/blackjack/index.tsx
@@ -10,6 +10,7 @@ import {
   BlackjackTemplate,
   GameStruct,
   GameType,
+  toDecimals,
 } from '@winrlabs/games';
 import {
   blackjackReaderAbi,
@@ -105,7 +106,11 @@ const defaultGameData = {
   payback: 0,
 };
 
+const MAX_WAGER_MULTIPLIER = 3;
+
 export default function BlackjackTemplateWithWeb3(props: TemplateWithWeb3Props) {
+  const maxWager = toDecimals((props?.maxWager || 100) / MAX_WAGER_MULTIPLIER, 2);
+
   const { gameAddresses, controllerAddress, cashierAddress, uiOperatorAddress, wagmiConfig } =
     useContractConfigContext();
 
@@ -762,16 +767,31 @@ export default function BlackjackTemplateWithWeb3(props: TemplateWithWeb3Props) 
     switch (gameEvent.program[0]?.type) {
       case BJ_EVENT_TYPES.Settled: {
         const status = Number(gameEvent.program[0].data.game.status);
+        const playerCardsEvent = gameEvent.program[2]?.data as BlackjackPlayerCardsEvent;
 
         // handle events by game status
         if (status == BlackjackGameStatus.FINISHED) {
-          console.log('game finished');
-          handleGameFinalizeEvent(gameEvent);
+          if (
+            gameEvent.program[2]?.type == 'PlayerCards' &&
+            playerCardsEvent.cards.amountCards == 2 &&
+            playerCardsEvent.cards.totalCount == 21 &&
+            activeMove == 'Created'
+          ) {
+            console.log('edge case');
+            setInitialDataFetched(true);
+            handleFirstBlackjackDistribution(gameEvent);
+            setTimeout(() => {
+              setInitialDataFetched(false);
+            }, 1000);
+          } else {
+            console.log('game finished');
+            handleGameFinalizeEvent(gameEvent);
+          }
         }
 
         console.log(activeMove, 'ACTIVE MOVE!');
         // handle events by active move
-        if (activeMove == 'Created') {
+        if (activeMove == 'Created' && playerCardsEvent.cards.totalCount !== 21) {
           setTimeout(() => {
             gameDataRead.refetch();
           }, 200);
@@ -827,6 +847,59 @@ export default function BlackjackTemplateWithWeb3(props: TemplateWithWeb3Props) 
   };
 
   // event handlers
+  const handleFirstBlackjackDistribution = (results: DecodedEvent<any, any>) => {
+    const settledEvent = results.program[0]?.data as BlackjackSettledEvent;
+    const playerCardsEvent = results.program[2]?.data as BlackjackPlayerCardsEvent;
+    const dealerCardsEvent = results.program[3]?.data as BlackjackPlayerCardsEvent;
+
+    const newPlayerHand = {
+      cards: {
+        cards: playerCardsEvent.cards.cards,
+        amountCards: playerCardsEvent.cards.cards.filter((n) => n !== 0).length,
+        totalCount: playerCardsEvent.cards.totalCount,
+        isSoftHand: playerCardsEvent.cards.isSoftHand,
+        canSplit: playerCardsEvent.cards?.canSplit || false,
+      },
+      hand: {
+        chipsAmount: formValues.firstHandWager,
+        isInsured: false,
+        status: BlackjackHandStatus.BLACKJACK,
+        isDouble: false,
+        isSplitted: false,
+        splittedHandIndex: null,
+      },
+      handId: playerCardsEvent.handIndex,
+    };
+
+    const newDealerHand = {
+      cards: {
+        cards: dealerCardsEvent.cards.cards,
+        amountCards: dealerCardsEvent.cards.cards.filter((n) => n !== 0).length,
+        totalCount: dealerCardsEvent.cards.totalCount,
+        isSoftHand: dealerCardsEvent.cards.isSoftHand,
+        canSplit: dealerCardsEvent.cards?.canSplit || false,
+      },
+      hand: {
+        chipsAmount: 0,
+        isInsured: false,
+        status: BlackjackHandStatus.BLACKJACK,
+        isDouble: false,
+        isSplitted: false,
+        splittedHandIndex: null,
+      },
+      handId: playerCardsEvent.handIndex,
+    };
+
+    setActiveGameHands((prev) => ({ ...prev, firstHand: newPlayerHand, dealer: newDealerHand }));
+
+    // set new game data
+    setActiveGameData((prev) => ({
+      ...prev,
+      activeHandIndex: Number(settledEvent.game.activeHandIndex),
+      status: Number(settledEvent.game.status),
+    }));
+  };
+
   const handlePlayerEvent = (results: DecodedEvent<any, any>) => {
     const hitCardEvent = results.program[0]?.data as BlackjackSettledEvent;
     const playerCardsEvent = results.program[2]?.data as BlackjackPlayerCardsEvent;
@@ -1288,7 +1361,7 @@ export default function BlackjackTemplateWithWeb3(props: TemplateWithWeb3Props) 
         activeGameHands={activeGameHands}
         initialDataFetched={initialDataFetched}
         minWager={props.minWager}
-        maxWager={props.maxWager}
+        maxWager={maxWager}
         onFormChange={(v) => setFormValues(v)}
         onGameCompleted={onGameCompleted}
         isControllerDisabled={isLoading}