+++ packages/web3-games/src/games/single-blackjack/index.tsx
@@ -11,6 +11,7 @@ import {
   SingleBJActiveGameHands,
   SingleBJDealFormFields,
   SingleBlackjackTemplate,
+  toDecimals,
 } from '@winrlabs/games';
 import {
   blackjackReaderAbi,
@@ -88,7 +89,11 @@ const defaultGameData = {
   payout: 0,
 };
 
+const MAX_WAGER_MULTIPLIER = 3;
+
 export default function SingleBlackjackGame(props: TemplateWithWeb3Props) {
+  const maxWager = toDecimals((props.maxWager || 100) / MAX_WAGER_MULTIPLIER, 2);
+
   const { gameAddresses, controllerAddress, cashierAddress, uiOperatorAddress, wagmiConfig } =
     useContractConfigContext();
 
@@ -1116,7 +1121,7 @@ export default function SingleBlackjackGame(props: TemplateWithWeb3Props) {
         activeGameHands={activeGameHands}
         initialDataFetched={initialDataFetched}
         minWager={props.minWager}
-        maxWager={props.maxWager}
+        maxWager={maxWager}
         onFormChange={(v) => setFormValues(v)}
         onGameCompleted={onGameCompleted}
         isControllerDisabled={isLoading}